Create module EnumForDistrictSelection
