+++ app/models/rent_farmland.rb
@@ -1,4 +1,6 @@
 class RentFarmland < ApplicationRecord
+  include EnumForDistrictSelection
+
   belongs_to :user
   default_scope -> { order(created_at: :desc) }
 end
